+++ packages/features/snapshots/commands.txt
@@ -6,9 +6,9 @@
  app generate extension @shopify/app           
  app generate schema    @shopify/app           
  app info               @shopify/app           
- app test event-topic   @shopify/app           
  auth logout            @shopify/cli           
  commands               @oclif/plugin-commands 
+ event trigger          @shopify/app           
  help                   @oclif/plugin-help     
  hydrogen add tailwind  @shopify/cli-hydrogen  
  hydrogen build         @shopify/cli-hydrogen
